refactor: Moved code to responsible for pointerEvents to its own hook (#62)
