+++ src/components/layout/index.jsx
@@ -1,12 +1,16 @@
 import React from "react";
 import PropTypes from "prop-types";
 
+import { useNavigationEffect } from "src/hooks/useNavigationEffect";
+
 import Header from "src/components/header";
 import Footer from "src/components/footer";
 
 import "./index.scss";
 
 const Layout = ({ children, isHome }) => {
+  useNavigationEffect();
+
   return (
     <>
       <Header isHome={isHome} />
